requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
